+++ test_opensearchpy/test_server/test_metrics.py
@@ -11,6 +11,8 @@ from __future__ import unicode_literals
 
 import time
 
+import pytest
+
 from opensearchpy import RequestsHttpConnection
 from opensearchpy.metrics.metrics_events import MetricsEvents
 from opensearchpy.metrics.metrics_none import MetricsNone
@@ -32,6 +34,12 @@ class TestMetrics(OpenSearchTestCase):
             client.indices.create(index=index_name, body=index_body)
         except Exception as e:
             assert False, f"Error creating index: {e}"
+
+    def test_metrics_raises_error_when_value_is_none(self) -> None:
+        # Test behavior when metrics is given None.
+        metrics = None
+        with pytest.raises(AttributeError):
+            get_client(metrics=metrics)
 
     def test_metrics_none_behavior(self) -> None:
         # Test behavior when metrics is an instance of MetricsNone